+++ tiger-test-lib/src/test/java/de/gematik/test/tiger/lib/integrationtest/TestContextFeatureTest.java
@@ -2,7 +2,7 @@
  * ${GEMATIK_COPYRIGHT_STATEMENT}
  */
 
-package drivers;
+package de.gematik.test.tiger.lib.integrationtest;
 
 import io.cucumber.junit.CucumberOptions;
 import net.serenitybdd.cucumber.CucumberWithSerenity;
@@ -10,8 +10,10 @@ import org.junit.runner.RunWith;
 
 @RunWith(CucumberWithSerenity.class)
 @CucumberOptions(
-    features = {"./src/test/resources/features/tiger-test-lib/testcontext.feature"},
+    features = {
+        "./src/test/resources/features/tiger-test-lib/testcontext.feature"
+    },
     plugin = {"json:target/cucumber-parallel/1.json"},
     glue = { "de.gematik.test.tiger.glue" })
-public class Parallel1IT {
+public class TestContextFeatureTest {
 }